Revert "Use simple Redis repository by default"
This reverts commit 8582b970

See gh-2122
