+++ spring-session-data-redis/src/main/java/org/springframework/session/data/redis/config/annotation/web/http/EnableRedisHttpSession.java
@@ -32,7 +32,6 @@ import org.springframework.session.Session;
 import org.springframework.session.SessionRepository;
 import org.springframework.session.config.annotation.web.http.EnableSpringHttpSession;
 import org.springframework.session.data.redis.RedisIndexedSessionRepository;
-import org.springframework.session.data.redis.RedisSessionRepository;
 import org.springframework.session.web.http.SessionRepositoryFilter;
 
 /**
@@ -54,8 +53,7 @@ import org.springframework.session.web.http.SessionRepositoryFilter;
  * }
  * </pre>
  *
- * More advanced configurations can extend {@link RedisHttpSessionConfiguration} or
- * {@link RedisIndexedHttpSessionConfiguration} instead.
+ * More advanced configurations can extend {@link RedisHttpSessionConfiguration} instead.
  *
  * @author Rob Winch
  * @author Vedran Pavic
@@ -65,7 +63,7 @@ import org.springframework.session.web.http.SessionRepositoryFilter;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 @Documented
-@Import(RedisHttpSessionConfigurationSelector.class)
+@Import(RedisHttpSessionConfiguration.class)
 @Configuration(proxyBeanMethods = false)
 public @interface EnableRedisHttpSession {
 
@@ -101,20 +99,18 @@ public @interface EnableRedisHttpSession {
 	FlushMode flushMode() default FlushMode.ON_SAVE;
 
 	/**
+	 * The cron expression for expired session cleanup job. By default runs every minute.
+	 * @return the session cleanup cron expression
+	 * @since 2.0.0
+	 */
+	String cleanupCron() default RedisHttpSessionConfiguration.DEFAULT_CLEANUP_CRON;
+
+	/**
 	 * Save mode for the session. The default is {@link SaveMode#ON_SET_ATTRIBUTE}, which
 	 * only saves changes made to session.
 	 * @return the save mode
 	 * @since 2.2.0
 	 */
 	SaveMode saveMode() default SaveMode.ON_SET_ATTRIBUTE;
-
-	/**
-	 * Indicate whether the {@link SessionRepository} should publish session events and
-	 * support fetching sessions by index. If true, a
-	 * {@link RedisIndexedSessionRepository} will be used in place of
-	 * {@link RedisSessionRepository}. This will result in slower performance.
-	 * @return true if indexing and events should be enabled, false otherwise
-	 */
-	boolean enableIndexingAndEvents() default false;
 
 }
