+++ spring-session-data-redis/src/test/java/org/springframework/session/data/redis/config/annotation/web/http/RedisHttpSessionConfigurationOverrideSessionTaskExecutors.java
@@ -39,17 +39,19 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.willAnswer;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 /**
  * @author Vladimir Tsanev
+ * @author Mark Paluch
  *
  */
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration
 @WebAppConfiguration
-class RedisIndexedHttpSessionConfigurationOverrideSessionTaskExecutor {
+class RedisHttpSessionConfigurationOverrideSessionTaskExecutors {
 
 	@Autowired
 	RedisMessageListenerContainer redisMessageListenerContainer;
@@ -57,17 +59,32 @@ class RedisIndexedHttpSessionConfigurationOverrideSessionTaskExecutor {
 	@Autowired
 	Executor springSessionRedisTaskExecutor;
 
+	@Autowired
+	Executor springSessionRedisSubscriptionExecutor;
+
 	@Test
-	void overrideSessionTaskExecutor() {
-		verify(this.springSessionRedisTaskExecutor, times(1)).execute(any(Runnable.class));
+	void overrideSessionTaskExecutors() {
+		verify(this.springSessionRedisSubscriptionExecutor, times(1)).execute(any(Runnable.class));
+		verify(this.springSessionRedisTaskExecutor, never()).execute(any(Runnable.class));
 	}
 
-	@EnableRedisHttpSession(enableIndexingAndEvents = true)
+	@EnableRedisHttpSession
 	@Configuration
 	static class Config {
 
 		@Bean
 		Executor springSessionRedisTaskExecutor() {
+			Executor executor = mock(Executor.class);
+			willAnswer((it) -> {
+				Runnable r = it.getArgument(0);
+				new Thread(r).start();
+				return null;
+			}).given(executor).execute(any());
+			return executor;
+		}
+
+		@Bean
+		Executor springSessionRedisSubscriptionExecutor() {
 			Executor executor = mock(Executor.class);
 			willAnswer((it) -> {
 				Runnable r = it.getArgument(0);
