+++ spring-session-docs/modules/ROOT/examples/resources/docs/http/HttpSessionListenerXmlTests-context.xml
@@ -14,7 +14,7 @@
 
 
 	<context:annotation-config/>
-	<bean class="org.springframework.session.data.redis.config.annotation.web.http.RedisIndexedHttpSessionConfiguration"/>
+	<bean class="org.springframework.session.data.redis.config.annotation.web.http.RedisHttpSessionConfiguration"/>
 
 	<bean class="docs.http.AbstractHttpSessionListenerTests"
 		factory-method="createMockRedisConnection"/>
